Main: Android - fold APKZip and APKFileSystem into standard headers
